+++ src/encoding/evm/strategy_encoder/group_swaps.rs
