+++ src/encoding/evm/strategy_encoder/mod.rs
@@ -1,4 +1,4 @@
+mod group_swaps;
 pub mod strategy_encoder_registry;
 mod strategy_encoders;
 mod strategy_validators;
-mod utils;
